feat: Add custom 404 page
Created a custom 404 page to enhance user experience for non-existent routes.
