+++ src/assets/raw/android-chrome-512x512.png
