+++ src/js/404.js
@@ -1,0 +1,2 @@
+import '../scss/404.scss';
+import "./analytics.js";
