+++ src/scss/_variables.scss
@@ -1,0 +1,15 @@
+$font-family: 'Poppins', 'Arial', sans-serif;
+$background-color: #D9E2EC;
+$text-color: #00509E;
+$primary-color: #4B5563;
+$hover-color: #3CB371;
+$link-color: #FFF;
+$error-color: #f44336;
+$border-radius: 4px;
+$gap: 16px;
+$button-padding: 24px;
+
+$font-size-large: 48px; // Large font size for main heading
+$font-size-medium: 24px; // Medium font size for subheading
+$font-size-small: 16px; // Small font size for additional text
+$font-size-link: 16px; // Font size for link text
