+++ src/views/404.ejs
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <title>404 - Page not found - Oops!</title>
+</head>
+<body>
+    <div class="error-container">
+        <header>
+            <h1>Page not found</h1>
+            <p>Oops! Looks like you’ve hit a dead end.</p>
+        </header>
+        <picture>
+            <source type="image/webp" srcset="../assets/images/confused-robot.png?as=webp">
+            <img src="../assets/images/confused-robot.png" alt="Confused robot looking for the page" width="300" height="270" />
+        </picture>
+        <div class="action">
+            <p>Don't worry, even robots get lost sometimes.</p>
+            <a href="/" class="home-link">Take Me Home</a>
+        </div>
+    </div>
+</body>
+</html>
