+++ src/views/index.ejs
@@ -12,7 +12,7 @@
     <meta property="og:type" content="website" />
     <meta property="og:title" content="Balázs Varga" />
     <meta property="og:description" content="Discover the professional profile of Balázs Varga. View links to LinkedIn, GitHub, Bluesky, and blog content." />
-    <meta property="og:image" content="../asset/resource/profile.png" />
+    <meta property="og:image" content="../assets/images/profile.png" />
     <meta property="og:image:alt" content="Balázs Varga's profile picture" />
 </head>
 <body>
@@ -20,8 +20,8 @@
         <div class="clazz-content">
             <header>
                 <picture>
-                    <source type="image/webp" srcset="../asset/resource/profile100w100h.png?as=webp" />
-                    <img width="100" height="100" src="../asset/resource/profile100w100h.png" alt="Balázs Varga" class="clazz-profile-image" />
+                    <source type="image/webp" srcset="../assets/images/profile100w100h.png?as=webp" />
+                    <img width="100" height="100" src="../assets/images/profile100w100h.png" alt="Balázs Varga" class="clazz-profile-image" />
                 </picture>
                 <h1 class="clazz-text-center">Balázs Varga</h1>
             </header>
